refactor: extract config, renderers, services, utils, routers from main
- config.py: GCP auth, env, Vertex client, templates
- renderers.py: HTML builders (table, summary, job result, equipment, e142 rows, errors)
- services/job_runner.py: csv_profile, csv_profile_no_header
- utils.py: single_line_message, exception_message, parse_float_or_none
- extractors/csv_utils.py: read_csv_dict_rows, read_csv_rows
- routers/pages.py: GET /, /me-check, /e-055, /e-251, /e-142, /area, etc.
- routers/downloads.py: GET /jobs/{job_id}/*.csv
- main.py: use above modules, remove inlined logic

Made-with: Cursor
